refactor: moved equipment logic in fighter class
