+++ core/views/sceneries/inventory_view.py
